+++ rollup.config.js
@@ -2,6 +2,8 @@ import fs from 'fs';
 import sourcemaps from 'rollup-plugin-sourcemaps';
 import {plugins} from './build/rollup_plugins';
 
+const version = JSON.parse(fs.readFileSync('package.json')).version;
+
 const production = process.env.BUILD === 'production';
 const outputFile = production ? 'dist/mapbox-gl.js' : 'dist/mapbox-gl-dev.js';
 
@@ -35,7 +37,8 @@ const config = [{
         format: 'umd',
         sourcemap: production ? true : 'inline',
         indent: false,
-        intro: fs.readFileSync(require.resolve('./rollup/bundle_prelude.js'), 'utf8')
+        intro: fs.readFileSync(require.resolve('./rollup/bundle_prelude.js'), 'utf8'),
+        banner: `/* Mapbox GL JS is licensed under the 3-Clause BSD License. Full text of license: https://github.com/mapbox/mapbox-gl-js/blob/v${version}/LICENSE.txt */`
     },
     treeshake: false,
     plugins: [