+++ src/ui/bind_handlers.js
@@ -21,7 +21,7 @@ const handlers = {
     touchZoomRotate
 };
 
-export default function bindHandlers(map: Map, options: {interactive: boolean}) {
+export default function bindHandlers(map: Map, options: {interactive: boolean, clickTolerance: number}) {
     const el = map.getCanvasContainer();
     let contextMenuEvent = null;
     let mouseDown = false;
@@ -152,7 +152,7 @@ export default function bindHandlers(map: Map, options: {interactive: boolean}) 
 
     function onClick(e: MouseEvent) {
         const pos = DOM.mousePos(el, e);
-        if (pos.equals(startPos)) {
+        if (pos.equals(startPos) || pos.dist(startPos) < options.clickTolerance) {
             map.fire(new MapMouseEvent('click', map, e));
         }
     }
@@ -182,6 +182,10 @@ export default function bindHandlers(map: Map, options: {interactive: boolean}) 
     }
 
     function onWheel(e: WheelEvent) {
+        if (options.interactive) {
+            map.stop();
+        }
+
         const mapEvent = new MapWheelEvent('wheel', map, e);
         map.fire(mapEvent);