+++ src/util/browser.js
@@ -1,6 +1,7 @@
 // @flow
 
 import window from './window';
+import type { Cancelable } from '../types/cancelable';
 
 const now = window.performance && window.performance.now ?
     window.performance.now.bind(window.performance) :
@@ -26,12 +27,9 @@ const exported = {
      */
     now,
 
-    frame(fn: Function) {
-        return raf(fn);
-    },
-
-    cancelFrame(id: number) {
-        return cancel(id);
+    frame(fn: Function): Cancelable {
+        const frame = raf(fn);
+        return { cancel: () => cancel(frame) };
     },
 
     getImageData(img: CanvasImageSource): ImageData {