+++ test/unit/source/raster_tile_source.test.js
@@ -129,6 +129,13 @@ test('RasterTileSource', (t) => {
         });
         window.server.respond();
     });
-    t.end();
 
+    t.test('cancels TileJSON request if removed', (t) => {
+        const source = createSource({ url: "/source.json" });
+        source.onRemove();
+        t.equal(window.server.lastRequest.aborted, true);
+        t.end();
+    });
+
+    t.end();
 });