+++ test/unit/ui/handler/box_zoom.test.js
@@ -4,9 +4,9 @@ import Map from '../../../../src/ui/map';
 import DOM from '../../../../src/util/dom';
 import simulate from 'mapbox-gl-js-test/simulate_interaction';
 
-function createMap(t) {
+function createMap(t, clickTolerance) {
     t.stub(Map.prototype, '_detectMissingCSS');
-    return new Map({ container: DOM.create('div', '', window.document.body) });
+    return new Map({ container: DOM.create('div', '', window.document.body), clickTolerance });
 }
 
 test('BoxZoomHandler fires boxzoomstart and boxzoomend events at appropriate times', (t) => {
@@ -129,6 +129,34 @@ test('BoxZoomHandler does not begin a box zoom on spurious mousemove events', (t
     simulate.mouseup(map.getCanvas(), {shiftKey: true, clientX: 0, clientY: 0});
     map._renderTaskQueue.run();
     t.equal(boxzoomstart.callCount, 0);
+    t.equal(boxzoomend.callCount, 0);
+
+    map.remove();
+    t.end();
+});
+
+test('BoxZoomHandler does not begin a box zoom until mouse move is larger than click tolerance', (t) => {
+    const map = createMap(t, 4);
+
+    const boxzoomstart = t.spy();
+    const boxzoomend   = t.spy();
+
+    map.on('boxzoomstart', boxzoomstart);
+    map.on('boxzoomend',   boxzoomend);
+
+    simulate.mousedown(map.getCanvas(), {shiftKey: true, clientX: 0, clientY: 0});
+    map._renderTaskQueue.run();
+    t.equal(boxzoomstart.callCount, 0);
+    t.equal(boxzoomend.callCount, 0);
+
+    simulate.mousemove(map.getCanvas(), {shiftKey: true, clientX: 3, clientY: 0});
+    map._renderTaskQueue.run();
+    t.equal(boxzoomstart.callCount, 0);
+    t.equal(boxzoomend.callCount, 0);
+
+    simulate.mousemove(map.getCanvas(), {shiftKey: true, clientX: 0, clientY: 4});
+    map._renderTaskQueue.run();
+    t.equal(boxzoomstart.callCount, 1);
     t.equal(boxzoomend.callCount, 0);
 
     map.remove();